+++ server/api/controllers/userController.js
@@ -1,23 +1,10 @@
+const _ = require('lodash');
 const userManager = require('../managers/userManager');
-const screenshotManager = require('../managers/screenshotManager');
-const cloudinaryService = require('../services/cloudinaryService');
 
 module.exports = {
-  getScores,
   getUser,
-  getSolvedScreenshots,
-  getAddedScreenshots,
-  getScreenshotRating,
+  updateUser,
 };
-
-async function getScores() {
-  const totalNbScreenshots = await screenshotManager.getTotalNb();
-  const scores = await userManager.getScores({ totalNbScreenshots });
-  return {
-    totalNbScreenshots,
-    scores,
-  };
-}
 
 async function getUser(req) {
   const { username } = req.user;
@@ -28,29 +15,31 @@ async function getUser(req) {
     email: user.email,
     nbSolvedScreenshots: user.solvedScreenshots,
     nbAddedScreenshots: user.addedScreenshots,
+    emailUpdates: user.emailUpdates,
   };
 }
 
-async function getSolvedScreenshots(req) {
+async function updateUser(req) {
   const { id } = req.user;
-  const screenshots = await userManager.getSolvedScreenshots(id);
-  return screenshots.map(screenshot => ({
-    ...screenshot,
-    imageUrl: cloudinaryService.pathToUrl(screenshot.imagePath),
-  }));
-}
+  const values = {};
 
-async function getAddedScreenshots(req) {
-  const { id } = req.user;
-  const screenshots = await userManager.getAddedScreenshots(id);
-  return screenshots.map(screenshot => ({
-    ...screenshot,
-    imageUrl: cloudinaryService.pathToUrl(screenshot.imagePath),
-  }));
-}
+  if (req.body.values.emailUpdates) {
+    const emailUpdatesValues = ['never', 'weekly', 'daily', 'asap'];
+    if (emailUpdatesValues.indexOf(req.body.values.emailUpdates) === -1) {
+      throw new Error(
+        `emailUpdates must be one of ${emailUpdatesValues.join(', ')}`
+      );
+    } else {
+      values.emailUpdates = req.body.values.emailUpdates;
+    }
+  }
 
-async function getScreenshotRating(req) {
-  const userId = req.user.id;
-  const { screenshotId } = req.body;
-  return userManager.getScreenshotRating({ screenshotId, userId });
+  if (_.isEmpty(values)) {
+    throw new Error('nothing to update');
+  }
+
+  await userManager.update(id, values);
+  return {
+    ok: true,
+  };
 }