+++ server/db/migrations/2019_07_07-1.sql
@@ -1,0 +1,10 @@
+ALTER TABLE `Users`
+  ADD `emailUpdates`
+    ENUM('asap','daily','weekly','never') NULL DEFAULT NULL AFTER `canModerateScreenshots`,
+  ADD `emailUpdateLastScreenshotId`
+    INT UNSIGNED NULL DEFAULT NULL AFTER `emailUpdates`;
+
+UPDATE `Users`
+  SET `Users`.`emailUpdateLastScreenshotId` = (
+    SELECT MAX(id) FROM `Screenshots` WHERE `Screenshots`.`approvalStatus` = 1
+  );