+++ src/App.js
@@ -9,6 +9,10 @@ import history from './history';
 
 import './App.css';
 
+import '../node_modules/noty/lib/noty.css';
+import '../node_modules/noty/lib/themes/sunset.css';
+import Noty from 'noty';
+
 import Layout from './components/Layout/Layout';
 
 // Basic components
@@ -27,6 +31,15 @@ import NewPasswordPage from './pages/login/NewPassword/NewPassword';
 
 // User components
 import UserPages from './pages/User/User';
+
+// Misc
+import EmailUpdatesUnsubscribe from './pages/misc/EmailUpdatesUnsubscribe/EmailUpdatesUnsubscribe';
+
+Noty.overrideDefaults({
+  theme: 'sunset',
+  timeout: 3000,
+  progressBar: false,
+});
 
 export default class App extends React.Component {
   componentDidMount() {
@@ -72,8 +85,12 @@ export default class App extends React.Component {
                 exact
                 component={EditScreenshotPage}
               />
-              <Route path="/la-fin" exact component={TheEnd} />
               <Route path="/moi" component={UserPages} />
+              <Route path="/la-fin" exact component={TheEnd} />
+              <Route
+                path="/email-updates/unsubscribe"
+                component={EmailUpdatesUnsubscribe}
+              />
               <Route component={NotFound} />
             </Switch>
           </Layout>