+++ src/services/userService.js
@@ -6,6 +6,7 @@ export default {
   fetchSolvedScreenshots,
   fetchAddedScreenshots,
   fetchScreenshotRating,
+  updateUser,
 };
 
 async function fetchScores() {
@@ -26,4 +27,8 @@ async function fetchAddedScreenshots() {
 
 async function fetchScreenshotRating({ screenshotId }) {
   return api.post('/user/screenshot-rating', { screenshotId });
+}
+
+async function updateUser(values) {
+  return api.post('/user/update', { values });
 }