+++ CHANGELOG.md
@@ -1,6 +1,18 @@
+# v9.0.0 - Crons!
+
+- Add crons
+- Add New Screenshot Email Updates
+- Add Noty (notifications in frontend)
+- Bug fixes
+
 ## v8.2.0
 
 - Remove service workers (they will be back)
+
+### v8.1.2
+
+- Fix password change
+- Better 404 page
 
 ### v8.1.1